Move SET functions to BaseUtils.kt
